tests/integration/ai_recruiter/requirements.txt
Faker==20.1.0
pandoc==2.3
pypdf
openai
pandas
chromadb
python-dotenv
fastapi
uvicorn
python-multipart
